Устранены замечания ревьюера
