--- src/historymanagers/Node.java
@@ -1,13 +1,0 @@
-package historymanagers;
-
-public class Node<E> {
-    E item;
-    Node<E> next;
-    Node<E> prev;
-
-    Node(Node<E> prev, E element, Node<E> next) {
-        this.item = element;
-        this.next = next;
-        this.prev = prev;
-    }
-}
